feat: Split cornerstoneHelper/useLoadImage into multiple hook files
