+++ packages/insight-viewer/src/hooks/useViewerLifecycle.ts
@@ -1,0 +1,16 @@
+import { useEffect } from 'react'
+import { init, dispose } from '../modules/cornerstoneHelper'
+
+export default function useViewerLifecycle(
+  ref: React.RefObject<HTMLDivElement>
+): void {
+  useEffect(() => {
+    if (!ref || !ref.current) return undefined
+    const element = ref.current
+    init(element)
+
+    return () => {
+      dispose(element)
+    }
+  }, [ref])
+}
