+++ packages/insight-viewer/src/modules/cornerstoneHelper/index.ts
@@ -6,4 +6,3 @@ export {
   loadImage,
   getDefaultViewportForImage
 } from './utils'
-export { useLoadImage } from './useLoadImage'
